+++ Tests/TestData/MultiFileCharacterization/VBToCSResults/ConvertWholeSolution/Prefix.VbLibrary/My Project/MyNamespace.Static.1.Designer.cs
@@ -355,7 +355,7 @@ namespace Prefix.VbLibrary.My
             [DebuggerHidden()]
             private void Dispose__Instance__<T>(ref T instance)
             {
-                instance = default;
+                instance = default(T);
             }
 
             [DebuggerHidden()]