+++ Tests/TestData/MultiFileCharacterization/VBToCSResults/ConvertWholeSolution/WindowsAppVb/My Project/MyNamespace.Static.1.Designer.cs
@@ -392,7 +392,7 @@ namespace WindowsAppVb.My
             [DebuggerHidden()]
             private void Dispose__Instance__<T>(ref T instance)
             {
-                instance = default;
+                instance = default(T);
             }
 
             [DebuggerHidden()]